Port the filepicker to the near present day
Make the filepicker work in a modern Android, and eliminate the
NDK symlink code in the process. The choices made in the filepicker
doesn't seem to actually be used anywhere yet, though. The filepicker
probably should be replaced with a more modern equivalent in the
future.

Also, there's a non-working attempt to make the YabbaDoo notification
show up in a recent Android.