Make lint a bit happier
This is mostly changes along the lines of "if it's a feature
of an API later than 8, don't support it". There are a few
bug fixes, too, but almost nothing more than "make it work
the way I believe it might have worked in 2014". The notification
on file error still isn't fixed.